add mvc

+++ src/main/java/com/example/mycoolapp/controller/EmployeeController.java
@@ -3,16 +3,20 @@ package com.example.mycoolapp.controller;
 import com.example.mycoolapp.entity.Employee;
 import com.example.mycoolapp.service.employee.EmployeeService;
 import org.springframework.web.bind.annotation.*;
+import tools.jackson.databind.json.JsonMapper;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api")
 public class EmployeeController {
+    private JsonMapper jsonMapper;
     private final EmployeeService employeeService;
 
-    public EmployeeController(EmployeeService employeeService) {
+    public EmployeeController(EmployeeService employeeService, JsonMapper jsonMapper) {
         this.employeeService = employeeService;
+        this.jsonMapper = jsonMapper;
     }
 
     @GetMapping("/employees")
@@ -31,17 +35,42 @@ public class EmployeeController {
     }
 
     @PostMapping("/employees")
-    public void save(@RequestBody Employee employee) {
-        this.employeeService.save(employee);
+    public Employee save(@RequestBody Employee employee) {
+       return this.employeeService.save(employee);
     }
 
     @PutMapping("/employees")
     public Employee update(@RequestBody Employee employee) {
-        return this.employeeService.update(employee);
+        return this.employeeService.save(employee);
+    }
+
+    @PatchMapping("/employees/{employeeId}")
+    public Employee pathEmployee(@PathVariable int employeeId, @RequestBody Map<String, Object> pathPayload) {
+        Employee employee = this.employeeService.findById(employeeId);
+
+        if (employee == null) {
+            throw new RuntimeException("Employee with id: " + employeeId + " not found");
+        }
+
+        if(pathPayload.containsKey("id")) {
+            throw new RuntimeException("Employee id not allowed in request body - " + employeeId);
+        }
+
+        Employee patchedEmployee = jsonMapper.updateValue(employee, pathPayload);
+
+        Employee dbEmployee = employeeService.save(patchedEmployee);
+
+        return dbEmployee;
     }
 
     @DeleteMapping("/employees/{employeeId}")
     public void deleteEmployee(@PathVariable int employeeId) {
+        Employee employee = this.employeeService.findById(employeeId);
+
+        if (employee == null) {
+            throw new RuntimeException("Employee with id: " + employeeId + " not found");
+        }
+        
         this.employeeService.deleteById(employeeId);
     }
 }
